requirements.txt
Flask==2.3.3
openai>=1.3.8
python-dotenv==1.0.0
werkzeug==2.3.7